src/test/java/com/kidaristudio/lecture/controller/LectureControllerTest.java
package com.kidaristudio.lecture.controller;

import com.kidaristudio.lecture.domain.Lecture;
import com.kidaristudio.lecture.service.LectureService;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.restdocs.mockmvc.MockMvcRestDocumentation;
import org.springframework.restdocs.operation.preprocess.Preprocessors;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultHandlers;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;

import java.time.LocalDateTime;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.BDDMockito.given;
import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;

import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@SuppressWarnings("SpringJavaInjectionPointsAutowiringInspection")
@SpringBootTest
@AutoConfigureMockMvc
@AutoConfigureRestDocs
//@WebMvcTest(LectureController.class)
class LectureControllerTest {

    @Autowired
    private MockMvc mockMvc;

//    @MockBean // LectureController 의존하는 빈을 모킹
//    private LectureService lectureService;

    @Autowired
    private LectureService lectureService;

    @DisplayName("강연 조회")
    @Test
    void getLectureById() throws Exception {

//        Lecture l = lectureService.getLectureById(2);


//        // given
//        given(lectureService.getLectureById(any()))
//                .willReturn(
//                        l
//                );

        // when & then
        mockMvc.perform(get("/lectures/1"))
                .andDo(print())
                .andDo(MockMvcRestDocumentation.document("lectures/getLectureById",
                        Preprocessors.preprocessRequest(prettyPrint()),
                        Preprocessors.preprocessResponse(prettyPrint())))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$.id").exists())
                .andExpect(jsonPath("$.id").value(1))
                .andExpect(jsonPath("$.speaker").value("강창진"))
                ;


//        mockMvc.perform(get("/lectures/1"))
//                .andExpect(status().isOk())
//                .andDo(print());
    }
}